#4 LocalDateTIme 사용을 위한 JavaTimeModule 의존성 추가, register Test 코드 작성 완료

+++ ex05/src/main/java/ecount/quotation/mapper/QuotationMapper.java
@@ -11,7 +11,7 @@ public interface QuotationMapper {
 
 	List<QuotationDTO> findAllQuot(); // 다중 조회
 
-	QuotationDTO findQuotByAccNum(Long acc_number); // 단일 조회
+	QuotationDTO findQuotByAccNum(int acc_number); // 단일 조회
 
 	int insertQuot(QuotationDTO quot);
